refactor: 파일명 및 구조 일부 수정
