+++ front/src/router/index.ts
@@ -1,13 +1,11 @@
 import { createRouter, createWebHistory } from "vue-router";
 import MainView from "../views/MainView.vue";
-import WriteView from "../views/board/WriteView.vue";
-import BoardListView from "../views/board/BoardListView.vue";
-import EditView from "../views/board/EditView.vue";
-import LoginView from "../views/user/LoginView.vue";
-import JoinView from "../views/user/JoinView.vue";
-import UserView from "../views/user/UserView.vue";
-import MemeView from "../views/meme/MemeMainView.vue";
-
+import WriteView from "../views/board/Write.vue";
+import BoardListView from "../views/board/List.vue";
+import EditView from "../views/board/Edit.vue";
+import LoginView from "../views/user/Login.vue";
+import JoinView from "../views/user/Join.vue";
+import UserView from "../views/user/User.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
