+++ front/src/views/board/Edit.vue
