+++ front/src/views/board/List.vue
