+++ front/src/views/board/Write.vue
