+++ front/src/views/meme/Meme.vue
